md-description-dl/requirements.txt
# Function dependencies, for example:
# package>=version
yt-dlp
google-cloud-storage
google-cloud-bigquery
google-cloud-firestore